requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
gunicorn==20.0.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
